requirements.txt
openai
PyPDF2
python-docx
colorama